Added corrected parsing of events.

+++ ClusteringApp/SeriousGameClustering/ClusteringModel.cs
@@ -16,6 +16,6 @@ namespace SeriousGameClustering
 		public Instigator MostOccuredHighInstigator { get; set; }
 		public char Gender { get; set; }
 		public int Age { get; set; }
-		public long DistanceTraveled { get; set; }
+		public double DistanceTraveled { get; set; }
 	}
 }
